1.0v su pirma strategija

+++ uzd-2/Galutinis_Mediana.cpp
@@ -4,7 +4,7 @@
 #include <deque>
 
 // VEKTORIAI
-double Galutinis_Mediana(std::vector<int>pazymiai, int egzaminas)
+double Galutinis_Mediana(std::vector<double>pazymiai, double egzaminas)
 {
 	double galutinis_mediana;
 	if (pazymiai.size() % 2 == 0)
@@ -22,7 +22,7 @@ double Galutinis_Mediana(std::vector<int>pazymiai, int egzaminas)
 	return galutinis_mediana;
 }
 // LIST
-double Galutinis_Mediana(std::list<int>pazymiai, int egzaminas)
+double Galutinis_Mediana(std::list<double>pazymiai, double egzaminas)
 {
 	double galutinis_mediana;
 	if (pazymiai.size() % 2 == 0)
@@ -30,8 +30,8 @@ double Galutinis_Mediana(std::list<int>pazymiai, int egzaminas)
 		pazymiai.push_back(egzaminas);
 		pazymiai.sort();
 
-		std::list<int>::iterator it = pazymiai.begin();
-		int x = pazymiai.size() / 2;
+		std::list<double>::iterator it = pazymiai.begin();
+		double x = pazymiai.size() / 2;
 		std::advance(it, x);
 		galutinis_mediana = *it;
 	}
@@ -40,10 +40,10 @@ double Galutinis_Mediana(std::list<int>pazymiai, int egzaminas)
 		pazymiai.push_back(egzaminas);
 		pazymiai.sort();
 
-		std::list<int>::iterator it1 = pazymiai.begin();
-		std::list<int>::iterator it2 = pazymiai.begin();
-		int x = (pazymiai.size() / 2) - 1;
-		int y = (pazymiai.size() / 2) + 1;
+		std::list<double>::iterator it1 = pazymiai.begin();
+		std::list<double>::iterator it2 = pazymiai.begin();
+		double x = (pazymiai.size() / 2) - 1;
+		double y = (pazymiai.size() / 2) + 1;
 
 		std::advance(it1, x);
 		std::advance(it2, y);
@@ -55,7 +55,7 @@ double Galutinis_Mediana(std::list<int>pazymiai, int egzaminas)
 	return galutinis_mediana;
 }
 // DEKAI
-double Galutinis_Mediana(std::deque<int>pazymiai, int egzaminas)
+double Galutinis_Mediana(std::deque<double>pazymiai, double egzaminas)
 {
 	double galutinis_mediana;
 	if (pazymiai.size() % 2 == 0)
